Move ocr3 events & errors to the public interface
