+++ chains/solana/contracts/programs/ccip-router/src/instructions/v1/ocr3base.rs
@@ -2,7 +2,8 @@ use anchor_lang::prelude::*;
 use anchor_lang::solana_program::sysvar;
 use anchor_lang::solana_program::{keccak, secp256k1_recover::*};
 
-use crate::{Ocr3Config, Ocr3ConfigInfo};
+use crate::ocr3base::{ConfigSet, Ocr3Error, Transmitted};
+use crate::state::{Ocr3Config, Ocr3ConfigInfo};
 
 #[constant]
 pub const MAX_ORACLES: usize = 16; // can set a maximum of 16 transmitters + 16 signers simultaneously in a single set config tx
@@ -229,54 +230,4 @@ fn assign_oracles<const A: usize>(oracles: &mut [[u8; A]], location: &mut [[u8; 
         location[i] = *o;
     }
     Ok(())
-}
-
-#[error_code]
-enum Ocr3Error {
-    #[msg("Invalid config: F must be positive")]
-    InvalidConfigFMustBePositive,
-    #[msg("Invalid config: Too many transmitters")]
-    InvalidConfigTooManyTransmitters,
-    #[msg("Invalid config: Too many signers")]
-    InvalidConfigTooManySigners,
-    #[msg("Invalid config: F is too high")]
-    InvalidConfigFIsTooHigh,
-    #[msg("Invalid config: Repeated oracle address")]
-    InvalidConfigRepeatedOracle,
-    #[msg("Wrong message length")]
-    WrongMessageLength,
-    #[msg("Config digest mismatch")]
-    ConfigDigestMismatch,
-    #[msg("Wrong number signatures")]
-    WrongNumberOfSignatures,
-    #[msg("Unauthorized transmitter")]
-    UnauthorizedTransmitter,
-    #[msg("Unauthorized signer")]
-    UnauthorizedSigner,
-    #[msg("Non unique signatures")]
-    NonUniqueSignatures,
-    #[msg("Oracle cannot be zero address")]
-    OracleCannotBeZeroAddress,
-    #[msg("Static config cannot be changed")]
-    StaticConfigCannotBeChanged,
-    #[msg("Incorrect plugin type")]
-    InvalidPluginType,
-    #[msg("Invalid signature")]
-    InvalidSignature,
-}
-
-#[event]
-pub struct ConfigSet {
-    pub ocr_plugin_type: u8,
-    pub config_digest: [u8; 32],
-    pub signers: Vec<[u8; 20]>,
-    pub transmitters: Vec<Pubkey>,
-    pub f: u8,
-}
-
-#[event]
-pub struct Transmitted {
-    pub ocr_plugin_type: u8,
-    pub config_digest: [u8; 32],
-    pub sequence_number: u64,
 }
